serparate folder for scripts created
